Adds regex to specify 16.+ versions of Darwin
Before this change, only one recent version of Darwin was correcly
having the compiler flags modified to allow for deprecated syscalls.
After this change, all versions of Darwin 16+ should correcly apply the
cxx flags. This is still a somewhat temporary fix because the real
issue (replacing the deprecated calls) has not been addressed.